pyproject.toml
[project]
name = "ecstasy"
version = "0.1.0"
description = "Benchmarking different foundational models in protein design"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "biopython>=1.85",
    "biotite>=1.3.0",
    "fire>=0.7.0",
    "matplotlib>=3.10.3",
    "memory-profiler>=0.61.0",
    "numpy>=2.3.0",
    "pandas>=2.3.0",
    "psutil>=7.0.0",
    "requests>=2.32.4",
    "seaborn>=0.13.2",
    "tmtools>=0.2.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
    "pytest>=8.4.1",
]
